fix(crane_msgs,crane_robot_receiver): kick_stateの×10スケールによるuint8オーバーフローを修正

+++ crane_robot_receiver/include/crane_robot_receiver/robot_feedback_protocol.hpp
@@ -178,7 +178,7 @@ struct RobotFeedback
 
   uint8_t counter = 0;
 
-  uint8_t kick_state = 0;
+  uint16_t kick_state = 0;
 
   uint8_t temperature[7] = {};
